+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.asurspace.whatisalertdialog_pl"
+        applicationId "com.asurspace.d_alert"
         minSdk 21
         targetSdk 32
         versionCode 2
