+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.asurspace.whatisalertdialog_pl">
+    package="com.asurspace.d_alert">
 
     <application
         android:allowBackup="true"
