+++ app/src/main/java/com/asurspace/d_alert/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.asurspace.whatisalertdialog_pl
+package com.asurspace.d_alert
 
 import android.graphics.Color
 import android.os.Bundle
@@ -9,11 +9,11 @@ import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.DrawableCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
-import com.asurspace.whatisalertdialog_pl.app_contract.ICustomToolbarTitleProvider
-import com.asurspace.whatisalertdialog_pl.app_contract.IProvideCustomAction
-import com.asurspace.whatisalertdialog_pl.app_contract.Navigator
-import com.asurspace.whatisalertdialog_pl.app_contract.model.CustomAction
-import com.asurspace.whatisalertdialog_pl.databinding.ActivityMainBinding
+import com.asurspace.d_alert.app_contract.ICustomToolbarTitleProvider
+import com.asurspace.d_alert.app_contract.IProvideCustomAction
+import com.asurspace.d_alert.app_contract.Navigator
+import com.asurspace.d_alert.app_contract.model.CustomAction
+import com.asurspace.d_alert.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity(), Navigator {
