+++ app/src/main/java/com/asurspace/d_alert/app_contract/IProvideCustomAction.kt
@@ -1,0 +1,7 @@
+package com.asurspace.d_alert.app_contract
+
+import com.asurspace.d_alert.app_contract.model.CustomAction
+
+interface IProvideCustomAction {
+    fun getCustomAction(): CustomAction
+}
