+++ app/src/main/java/com/asurspace/d_alert/dialog_fragments/EditableDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.asurspace.whatisalertdialog_pl.dialog_fragments
+package com.asurspace.d_alert.dialog_fragments
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -10,15 +10,14 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.LifecycleOwner
-import com.asurspace.whatisalertdialog_pl.R
-import com.asurspace.whatisalertdialog_pl.databinding.CustomEditablieItemBinding
+import com.asurspace.d_alert.R
+import com.asurspace.d_alert.databinding.CustomEditablieItemBinding
 
-typealias EditableReusableDialogResultListener = (requestKey: String, volume: Int) -> Unit
+typealias EditableDialogResultListener = (Int) -> Unit
 
-class EditableReusableDialogFragment : DialogFragment() {
+class EditableDialogFragment : DialogFragment() {
 
     private val receivedVolumeValue get() = requireArguments().getInt(ARG_VALUE)
-    private val requestKey get() = requireArguments().getString(ARG_REQUEST_KEY)!!
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val editableBinding = CustomEditablieItemBinding.inflate(layoutInflater)
@@ -63,8 +62,12 @@ class EditableReusableDialogFragment : DialogFragment() {
         Log.d(TAG, "onCancel")
     }
 
-    private fun setFragmentResult(response: Int) { parentFragmentManager
-        .setFragmentResult(requestKey, bundleOf(KEY_VOLUME_RESPONSE to response)) }
+    private fun setFragmentResult(response: Int) {
+        parentFragmentManager.setFragmentResult(
+            REQUEST_KEY,
+            bundleOf(KEY_VOLUME_RESPONSE to response)
+        )
+    }
 
     companion object {
         @JvmStatic
@@ -77,28 +80,21 @@ class EditableReusableDialogFragment : DialogFragment() {
         private val ARG_VALUE = "ARG_VALUE"
 
         @JvmStatic
-        private val ARG_REQUEST_KEY = "ARG_REQUEST_KEY"
+        val REQUEST_KEY = "$TAG:defaultRequestKey"
 
-        @JvmStatic
-        val DEFAULT_REQUEST_KEY = "$TAG:defaultRequestKey"
-
-        fun show(fragmentManager: FragmentManager, requestKey: String, volumeValue: Int) {
+        fun show(fragmentManager: FragmentManager, color: Int) {
             val editableDialogFragment = EditableReusableDialogFragment()
-            editableDialogFragment.arguments = bundleOf(
-                ARG_REQUEST_KEY to requestKey,
-                ARG_VALUE to volumeValue
-            )
+            editableDialogFragment.arguments = bundleOf(ARG_VALUE to color)
             editableDialogFragment.show(fragmentManager, TAG)
         }
 
         fun setUpListener(
             fragmentManager: FragmentManager,
             lifecycleOwner: LifecycleOwner,
-            requestKey: String,
-            listener: EditableReusableDialogResultListener
+            listener: EditableDialogResultListener
         ) {
-            fragmentManager.setFragmentResultListener(requestKey, lifecycleOwner) { key, result ->
-                listener.invoke(key, result.getInt(KEY_VOLUME_RESPONSE))
+            fragmentManager.setFragmentResultListener(REQUEST_KEY, lifecycleOwner) { _, result ->
+                listener.invoke(result.getInt(KEY_VOLUME_RESPONSE))
             }
         }
     }
