+++ app/src/main/java/com/asurspace/d_alert/dialog_fragments/EditableReuseableDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.asurspace.whatisalertdialog_pl.dialog_fragments
+package com.asurspace.d_alert.dialog_fragments
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -10,14 +10,15 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.LifecycleOwner
-import com.asurspace.whatisalertdialog_pl.R
-import com.asurspace.whatisalertdialog_pl.databinding.CustomEditablieItemBinding
+import com.asurspace.d_alert.R
+import com.asurspace.d_alert.databinding.CustomEditablieItemBinding
 
-typealias EditableDialogResultListener = (Int) -> Unit
+typealias EditableReusableDialogResultListener = (requestKey: String, volume: Int) -> Unit
 
-class EditableDialogFragment : DialogFragment() {
+class EditableReusableDialogFragment : DialogFragment() {
 
     private val receivedVolumeValue get() = requireArguments().getInt(ARG_VALUE)
+    private val requestKey get() = requireArguments().getString(ARG_REQUEST_KEY)!!
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val editableBinding = CustomEditablieItemBinding.inflate(layoutInflater)
@@ -62,12 +63,8 @@ class EditableDialogFragment : DialogFragment() {
         Log.d(TAG, "onCancel")
     }
 
-    private fun setFragmentResult(response: Int) {
-        parentFragmentManager.setFragmentResult(
-            REQUEST_KEY,
-            bundleOf(KEY_VOLUME_RESPONSE to response)
-        )
-    }
+    private fun setFragmentResult(response: Int) { parentFragmentManager
+        .setFragmentResult(requestKey, bundleOf(KEY_VOLUME_RESPONSE to response)) }
 
     companion object {
         @JvmStatic
@@ -80,21 +77,28 @@ class EditableDialogFragment : DialogFragment() {
         private val ARG_VALUE = "ARG_VALUE"
 
         @JvmStatic
-        val REQUEST_KEY = "$TAG:defaultRequestKey"
+        private val ARG_REQUEST_KEY = "ARG_REQUEST_KEY"
 
-        fun show(fragmentManager: FragmentManager, color: Int) {
+        @JvmStatic
+        val DEFAULT_REQUEST_KEY = "$TAG:defaultRequestKey"
+
+        fun show(fragmentManager: FragmentManager, requestKey: String, volumeValue: Int) {
             val editableDialogFragment = EditableReusableDialogFragment()
-            editableDialogFragment.arguments = bundleOf(ARG_VALUE to color)
+            editableDialogFragment.arguments = bundleOf(
+                ARG_REQUEST_KEY to requestKey,
+                ARG_VALUE to volumeValue
+            )
             editableDialogFragment.show(fragmentManager, TAG)
         }
 
         fun setUpListener(
             fragmentManager: FragmentManager,
             lifecycleOwner: LifecycleOwner,
-            listener: EditableDialogResultListener
+            requestKey: String,
+            listener: EditableReusableDialogResultListener
         ) {
-            fragmentManager.setFragmentResultListener(REQUEST_KEY, lifecycleOwner) { _, result ->
-                listener.invoke(result.getInt(KEY_VOLUME_RESPONSE))
+            fragmentManager.setFragmentResultListener(requestKey, lifecycleOwner) { key, result ->
+                listener.invoke(key, result.getInt(KEY_VOLUME_RESPONSE))
             }
         }
     }
