--- app/src/main/java/com/asurspace/whatisalertdialog_pl/app_contract/IProvideCustomAction.kt
@@ -1,7 +1,0 @@
-package com.asurspace.whatisalertdialog_pl.app_contract
-
-import com.asurspace.whatisalertdialog_pl.app_contract.model.CustomAction
-
-interface IProvideCustomAction {
-    fun getCustomAction(): CustomAction
-}
