+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "WhatIsAlertDialog_PL"
+rootProject.name = "D_Alert"
 include ':app'
